# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: TopTrader API
  version: "1.0"
  description: >
    REST API for the TopTrader stock trading simulator. Session-based auth
    (ADR 0004) — the SESSION cookie is set on register/login and must be sent
    (withCredentials) on every subsequent request. All errors use RFC 7807
    application/problem+json.
servers:
  - url: https://api.toptrader.example
    description: Production (placeholder domain, see ADR 0005)
security:
  - sessionAuth: []
paths:
...
  /trades/buy:
    post:
      summary: Buy whole shares (US-5)
      description: >
        Price is never client-supplied — the server fetches the current quote
        at the moment this request is processed and executes at that price.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeRequest"
      responses:
        "201":
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          description: Insufficient cash, or zero/negative/fractional quantity
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Unknown ticker
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /trades/sell:
    post:
      summary: Sell whole shares held (US-6)
      description: >
        Price is never client-supplied — the server fetches the current quote
        at the moment this request is processed and executes at that price.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeRequest"
      responses:
        "201":
          description: Trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          description: Quantity greater than shares held, or zero/negative/fractional quantity
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Unknown ticker, or no holding in that ticker
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /trades/holdings/{ticker}:
    get:
      summary: Look up the logged-in user's holding in one ticker (US-6)
      description: >
        Used to conditionally show the sell form only when the user actually
        holds shares of this ticker. Cash balance is not included here — see
        GET /auth/session.
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
          example: AAPL
      responses:
        "200":
          description: Current holding in this ticker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
        "404":
          description: Unknown ticker, or no holding in that ticker
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    ProblemDetail:
      type: object
      description: RFC 7807 problem detail (Spring Boot's built-in ProblemDetail type)
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
...
    TradeRequest:
      type: object
      required: [ticker, quantity]
      properties:
        ticker:
          type: string
        quantity:
          type: integer
          minimum: 1
    Holding:
      type: object
      properties:
        ticker:
          type: string
        quantity:
          type: integer
        averageCostBasis:
          type: number
          format: double
        currentPrice:
          type: number
          format: double
        marketValue:
          type: number
          format: double
        unrealizedGainLoss:
          type: number
          format: double
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ticker:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: integer
        pricePerShare:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        executedAt:
          type: string
          format: date-time
    TradeResult:
      type: object
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        cashBalance:
          type: number
          format: double
        holding:
          $ref: "#/components/schemas/Holding"
          nullable: true
          description: Null when a sell fully liquidates the position